+++ include/grpcpp/security/credentials.h
@@ -130,6 +130,15 @@ class CallCredentials : private grpc::GrpcLibraryCodegen {
   virtual SecureCallCredentials* AsSecureCredentials() = 0;
 };
 
+// The range of TLS versions that may be negotiated during the handshake.
+struct TlsVersionRange {
+  TlsVersionRange() : min_tls_version(TLS1_2), max_tls_version(TLS1_3) {}
+  TlsVersionRange(grpc_tls_version min_version, grpc_tls_version max_version)
+      : min_tls_version(min_version), max_tls_version(max_version) {}
+  grpc_tls_version min_tls_version;
+  grpc_tls_version max_tls_version;
+};
+
 /// Options used to build SslCredentials.
 struct SslCredentialsOptions {
   /// The buffer containing the PEM encoding of the server root certificates. If
@@ -147,6 +156,9 @@ struct SslCredentialsOptions {
   /// This parameter can be empty if the client does not have a certificate
   /// chain.
   grpc::string pem_cert_chain;
+
+  /// The range of TLS versions allowed by the client.
+  TlsVersionRange tls_version_range;
 };
 
 // Factories for building different types of Credentials The functions may
